Backend/modelnew/fixed_requirements.txt
absl-py==2.1.0
accelerate==1.1.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.20
albumentations==1.4.21
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
av==13.1.0
beautifulsoup4==4.12.3
black==24.10.0
boto3==1.35.58
botocore==1.35.58
Brotli==1.0.9
cachetools==5.5.0
cchardet==2.1.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chumpy==0.70
cityscapesScripts==2.2.4
click==8.1.7
clip==1.0
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cupy==13.3.0
cycler==0.12.1
Cython==3.0.11
darkdetect==0.7.1
datasets==3.2.0
debugpy==1.8.7
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
detectron2==0.6                  
diffusers==0.31.0
dill==0.3.8
dm-sonnet==1.23
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.4
fast-soft-sort==0.1
fastapi==0.115.0
fastrlock==0.8.2
ffmpy==0.4.0
filelock==3.14.0
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.54.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
groundingdino==0.1.0                
groundingdino-py==0.4.0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
hickle==5.0.3
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.36.0
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.9
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
json-tricks==3.17.3
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.11.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl_fft==1.3.10
mkl_random==1.2.7
mkl-service==2.4.0
ml-dtypes==0.4.1
mmcv-full==1.7.1
mmengine==0.10.5
mmpose==0.28.0
mmpt==0.0.1                
model-index==0.1.11
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
munkres==1.1.4
mypy-extensions==1.0.0
namex==0.0.8
nest_asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.0
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.82
nvitop==1.4.0
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
orjson==3.10.7
oss2==2.17.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
pexpect==4.9.0
pickle5==0.0.11
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
plyfile==1.1
portalocker==3.0.0
prettytable==3.11.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyglet==1.5.27
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.1.4
pyqtdarktheme==2.1.0
pyquaternion==0.9.9
pyrender==0.1.45
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytorch-ignite==0.2.0
pytorch-metric-learning==2.6.0
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9
ruff==0.6.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.17.0
setproctitle==1.3.3
setuptools==60.2.0
shapely==2.0.6
shellingham==1.5.4
shiboken6==6.7.3
simplejson==3.19.3
simsimd==6.0.1
six==1.16.0
smmap==5.0.1
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.2
starlette==0.38.6
stringzilla==3.10.5
supervision==0.6.0
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
termcolor==2.5.0
tf_keras==2.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
timesformer==1.0                  
timm==0.4.12
tokenizers==0.20.0
tomli==2.0.2
tomlkit==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchgeometry==0.1.2
torchvideotransforms==0.1.2
torchvision==0.12.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.2
trimesh==4.5.2
typer==0.12.5
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.31.0
vit-pytorch==1.9.0
wandb==0.18.5
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
wget==3.2
wheel==0.44.0
wrapt==1.17.0
xtcocotools==1.14.3
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.2
yarl==1.18.3
zipp==3.20.2
